Move logic from handler into action
* create action package
* move Channel into action
* move SlackAPI into action
* handler.ServeHTTP now only has to care about pulling values out of
  the r.PostForm, creating an action with action.New(), and performing
  the action

The tests for each action are still being driven through handler at
present. I'm going to leave these for now because I don't want to create
a provider (e.g. ActionProvider/ActionFactory) for the purposes of
faking it out in handler, even though it would make the ServeHTTP tests
as simple as the code in ServeHTTP now is.